config: Add TOML config
This patch adds a new type of config file for mixer, based on the TOML
format. The new format is not compatible with the old parser and
vice-versa. The new flag --new-config enables the use of the new format.

This patch also adds to the parser the properties that exist in the
config file but were not parsed by the older parser.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>