config: remove buildBundlesConfig
Builder config object already contains all the properties of
buildBundlesConfig, so it can be used instead. Also, for new config,
those values have already been parsed with the rest of the config file,
so there is no need to parse them again.

Since the config also has the default values for the server section,
there is no difference in adding the server section to the server.ini
file or not adding it and use the defaults in swupd/config.go, since
they are the same. This allows the check for 'HasServerSection'to be
removed.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>